+++ src/component/Rigidbody.java
@@ -7,7 +7,7 @@ import physics.Physics;
 public class Rigidbody extends Component{
 
 	public Vector2 velocity = Vector2.ZERO;
-	public double mass = 1;
+	public float mass = 1;
 	
 	public Vector2 oldVeclocity = Vector2.ZERO;
