+++ src/game/scene/Scene_1.java
@@ -19,8 +19,8 @@ public class Scene_1 extends Scene {
 		Gizmo.isEnable = true;
 		
 		for(int i = 0; i < 2000; i++){
-			CellObject c1 = new CellObject(Math.random()+2, Color.RED);
-			c1.transform.position = new Vector2(Math.random() * Renderer.WIDTH - Renderer.WIDTH/2, Math.random() * Renderer.HEIGHT - Renderer.HEIGHT/2);
+			CellObject c1 = new CellObject((float)(Math.random()+2.0f), Color.RED);
+			c1.transform.position = new Vector2((float)(Math.random() * Renderer.WIDTH - Renderer.WIDTH/2.0f), (float)(Math.random() * Renderer.HEIGHT - Renderer.HEIGHT/2.0f));
 		}
 		
 	}
